+++ tests/test_oas/test_oas.py
@@ -229,20 +229,17 @@ def test_ep_without_ret():
     get_resp_schemas(ep, {}, "")
 
 
-# def test_ep_with_auth():
+def test_ep_with_auth():
 
-#     async def get_user(
-#         token: Annotated[str, OAuth2PasswordFlow(token_url="token")],
-#     ): ...
+    async def get_user(token: str): ...
 
-#     route = Route()
-#     route.get(get_user)
+    route = Route()
+    route.get(auth_scheme=OAuth2PasswordFlow(token_url="token"))(get_user)
 
-#     ep = route.get_endpoint("GET")
-#     ep.setup()
+    ep = route.get_endpoint("GET")
+    ep.setup()
 
-#     token = ep.sig.plugins["token"]
 
-#     sc = {}
-#     get_ep_security(ep, sc)
-#     assert sc["OAuth2PasswordBearer"]
+    sc = {}
+    get_ep_security(ep, sc)
+    assert sc["OAuth2PasswordBearer"]