+++ CHANGELOG.md
@@ -577,3 +577,62 @@ res = await lc.submit_form(
 
 
 ## version 0.1.15
+
+
+### Fix
+
+- fix a bug where if `lifespan` is not provided and app starup fails, it will fail silently.
+
+- fix a bug where if endponit returns non-stream response, media-type would be lost, no will include the media-type of first response.
+
+
+
+### Features
+
+- automatic generate oauth2-compatible access token based on user returns
+
+```python
+from lihil import Lihil, Payload, Route, field
+from lihil.config import AppConfig, SecurityConfig
+from lihil.plugins.auth.jwt import JWToken, JWTPayload
+from lihil.plugins.auth.oauth import OAuth2PasswordFlow, OAuthLoginForm
+
+me = Route("me")
+token = Route("token")
+
+
+class UserPayload(JWTPayload):
+    __jwt_claims__ = {"expires_in": 300}
+
+    user_id: str = field(name="sub")
+
+
+class User(Payload):
+    name: str
+    email: str
+
+
+@me.get(auth_scheme=OAuth2PasswordFlow(token_url="token"))
+async def get_user(token: JWToken[UserPayload]) -> User:
+    assert token.user_id == "user123"
+    return User(name="user", email="user@email.com")
+
+
+@token.post
+async def create_token(credentials: OAuthLoginForm) -> JWToken[UserPayload]:
+    assert credentials.username == "admin" and credentials.password == "admin"
+    return UserPayload(user_id="user123")
+
+
+lhl = Lihil[None](
+    routes=[users, token],
+    app_config=AppConfig(
+        security=SecurityConfig(jwt_secret="mysecret", jwt_algorithms=["HS256"])
+    ),
+)
+```
+
+
+### Improvements
+
+- No longer automatically drops resonse body when status code < 200 or in (204, 205, 304), instead, user should declare its return with `lihil.Empty`